+++ uniCloud-aliyun/cloudfunctions/router/service/user/kh/getWeixinMPurl.js
@@ -1,8 +1,8 @@
 'use strict';
 module.exports = {
 	/**
-	 * 获取小程序scheme码，适用于短信、邮件、外部网页等拉起小程序的业务场景。通过该接口，可以选择生成到期失效和永久有效的小程序码，目前仅针对国内非个人主体的小程序开放
-	 * @url user/kh/getWeixinMPscheme 前端调用的url参数地址
+	 * 获取小程序 URL Link，适用于短信、邮件、网页、微信内等拉起小程序的业务场景。通过该接口，可以选择生成到期失效和永久有效的小程序链接，有数量限制，目前仅针对国内非个人主体的小程序开放
+	 * @url user/kh/getWeixinMPurl 前端调用的url参数地址
 	 * data 请求参数 说明
 	 * @param {String} path    小程序页面路径
 	 * @param {String} query   小程序页面参数
@@ -23,13 +23,10 @@ module.exports = {
 			env_version
 		} = data;
 		// 超过180天有10万个限制,故这里强制设为90天
-		res = await vk.openapi.weixin.urlscheme.generate({
-			// 跳转到的目标小程序信息。
-			jump_wxa: {
-				path, // 小程序页面路径
-				query, // 小程序页面参数
-				env_version
-			},
+		res = await vk.openapi.weixin.urllink.generate({
+			path,
+			query,
+			env_version,
 			is_expire: true,
 			expire_type: 1,
 			expire_interval: 90, // 90天后失效